# backend/openapi.yaml
openapi: 3.1.1
info:
  description: REST API for the FlowForm survey platform.
  contact:
    name: FlowForm Support
    email: support@flow-form.com.au
  title: FlowForm API
  version: 0.1.0
servers:
- url: http://localhost:5000
  description: Local development
security:
- BearerAuth: []
...
paths:
...
  /api/v1/projects/{project_id}/roles:
    get:
      operationId: listRoles
      summary: List project roles
      description: List project roles
      tags:
      - Roles
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createRole
      summary: Create project role
      description: Create project role
      tags:
      - Roles
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRoleRequest'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{project_id}/roles/{role_id}:
    patch:
      operationId: updateRole
      summary: Update project role
      description: Update project role
      tags:
      - Roles
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRoleRequest'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteRole
      summary: Delete project role
      description: Delete project role
      tags:
      - Roles
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      responses:
        '200':
          description: Successful response.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/projects/{project_id}/surveys/{survey_id}/members:
    get:
      operationId: listSurveyMembers
      summary: List survey member role assignments
      description: List survey member role assignments
      tags:
      - Survey Members
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: survey_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SurveyMemberRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: assignSurveyMemberRole
      summary: Assign survey role to member
      description: Assign survey role to member
      tags:
      - Survey Members
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: survey_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignSurveyMemberRoleRequest'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyMemberRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{project_id}/surveys/{survey_id}/members/{membership_id}:
    patch:
      operationId: updateSurveyMemberRole
      summary: Update survey member role assignment
      description: Update survey member role assignment
      tags:
      - Survey Members
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: survey_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSurveyMemberRoleRequest'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyMemberRoleResponses'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: removeSurveyMemberRole
      summary: Remove survey member role assignment
      description: Remove survey member role assignment
      tags:
      - Survey Members
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: survey_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      - name: membership_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
          maximum: 2147483647
      responses:
        '200':
          description: Successful response.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Machine-readable error code.
          maxLength: 128
        message:
          type: string
          description: Human-readable error message.
          maxLength: 1000
        details:
          type: object
          description: Optional structured extras. Pydantic validation errors place their per-field failures here as ``details.errors``.
          additionalProperties: true
...
    ProjectRoleResponses:
      description: API response shape for a project role.
      properties:
        id:
          title: Id
          type: integer
        project_id:
          title: Project Id
          type: integer
        name:
          title: Name
          type: string
          maxLength: 120
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        is_system_role:
          title: Is System Role
          type: boolean
        permissions:
          items:
            enum:
            - project:edit
            - project:delete
            - project:manage_members
            - project:manage_roles
            - survey:view
            - survey:create
            - survey:edit
            - survey:delete
            - survey:publish
            - survey:archive
            - submission:view
            type: string
            maxLength: 22
          maxItems: 11
          title: Permissions
          type: array
        created_at:
          format: date-time
          title: Created At
          type: string
          maxLength: 35
      required:
      - id
      - project_id
      - name
      - description
      - is_system_role
      - permissions
      - created_at
      title: ProjectRoleResponses
      type: object
    CreateProjectRoleRequest:
      description: Request body for creating a new project role.
      properties:
        name:
          maxLength: 80
          minLength: 1
          title: Name
          type: string
        description:
          anyOf:
          - maxLength: 500
            minLength: 1
            type: string
          - type: 'null'
          default: null
          title: Description
        permissions:
          items:
            enum:
            - project:edit
            - project:delete
            - project:manage_members
            - project:manage_roles
            - survey:view
            - survey:create
            - survey:edit
            - survey:delete
            - survey:publish
            - survey:archive
            - submission:view
            type: string
            maxLength: 22
          maxItems: 11
          title: Permissions
          type: array
          uniqueItems: true
      required:
      - name
      title: CreateProjectRoleRequest
      type: object
    UpdateProjectRoleRequest:
      description: Request body for updating a project role.
      properties:
        name:
          anyOf:
          - maxLength: 80
            minLength: 1
            type: string
          - type: 'null'
          default: null
          title: Name
        description:
          anyOf:
          - maxLength: 500
            minLength: 1
            type: string
          - type: 'null'
          default: null
          title: Description
        permissions:
          anyOf:
          - items:
              enum:
              - project:edit
              - project:delete
              - project:manage_members
              - project:manage_roles
              - survey:view
              - survey:create
              - survey:edit
              - survey:delete
              - survey:publish
              - survey:archive
              - submission:view
              type: string
              maxLength: 22
            maxItems: 11
            type: array
            uniqueItems: true
          - type: 'null'
          default: null
          title: Permissions
      title: UpdateProjectRoleRequest
      type: object
...
    SurveyMemberRoleResponses:
      description: API response shape for a survey membership role assignment.
      properties:
        project_id:
          title: Project Id
          type: integer
        survey_id:
          title: Survey Id
          type: integer
        membership_id:
          title: Membership Id
          type: integer
        role_id:
          title: Role Id
          type: integer
        created_at:
          format: date-time
          title: Created At
          type: string
          maxLength: 35
        member:
          anyOf:
          - $ref: '#/components/schemas/SurveyMemberResponses'
          - type: 'null'
          default: null
        role:
          anyOf:
          - $ref: '#/components/schemas/SurveyRoleResponses'
          - type: 'null'
          default: null
      required:
      - project_id
      - survey_id
      - membership_id
      - role_id
      - created_at
      title: SurveyMemberRoleResponses
      type: object
    AssignSurveyMemberRoleRequest:
      description: Request body for assigning a survey role to a project member for a survey.
      properties:
        membership_id:
          maximum: 2147483647
          minimum: 1
          title: Membership Id
          type: integer
        role_id:
          maximum: 2147483647
          minimum: 1
          title: Role Id
          type: integer
      required:
      - membership_id
      - role_id
      title: AssignSurveyMemberRoleRequest
      type: object
    UpdateSurveyMemberRoleRequest:
      description: Request body for updating a survey member role assignment.
      properties:
        role_id:
          maximum: 2147483647
          minimum: 1
          title: Role Id
          type: integer
      required:
      - role_id
      title: UpdateSurveyMemberRoleRequest
      type: object
...
  responses:
    BadRequestError:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_email_claim:
              summary: MissingEmailClaimError
              value:
                code: MISSING_EMAIL_CLAIM
                message: ID token did not contain an email claim.
            survey_no_response_store:
              summary: SurveyNoResponseStoreError
              value:
                code: INVALID_REQUEST
                message: <example>
            submission_invalid:
              summary: SubmissionInvalidError
              value:
                code: INVALID_SUBMISSION
                message: Submission is invalid.
            submission_invalid_timestamps:
              summary: SubmissionInvalidTimestampsError
              value:
                code: INVALID_SUBMISSION_TIMESTAMPS
                message: Submission timestamps are invalid.
            submission_answers_required:
              summary: SubmissionAnswersRequiredError
              value:
                code: SUBMISSION_ANSWERS_REQUIRED
                message: At least one answer is required.
    UnauthorizedError:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_id_token_subject:
              summary: InvalidIdTokenSubjectError
              value:
                code: INVALID_ID_TOKEN
                message: ID token did not contain a valid subject.
            token_subject_mismatch:
              summary: TokenSubjectMismatchError
              value:
                code: TOKEN_SUBJECT_MISMATCH
                message: ID token subject did not match the access token subject.
            link_auth_required:
              summary: LinkAuthRequiredError
              value:
                code: LINK_AUTH_REQUIRED
                message: This link requires authentication.
    ForbiddenError:
      description: Authenticated but not authorized for this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: ForbiddenError
              value:
                code: FORBIDDEN
                message: You do not have permission to access this resource.
            link_inactive:
              summary: LinkInactiveError
              value:
                code: LINK_INACTIVE
                message: This link is inactive
            link_expired:
              summary: LinkExpiredError
              value:
                code: LINK_EXPIRED
                message: This link has expired
            link_no_response:
              summary: LinkNoResponseError
              value:
                code: LINK_NO_RESPONSE
                message: This link does not allow submissions
            link_assignment_mismatch:
              summary: LinkAssignmentMismatchError
              value:
                code: LINK_ASSIGNED_TO_ANOTHER_USER
                message: This link is assigned to a different user.
            link_already_used:
              summary: LinkAlreadyUsedError
              value:
                code: LINK_ALREADY_USED
                message: This link has already been used.
            survey_not_accessible:
              summary: SurveyNotAccessibleError
              value:
                code: SURVEY_NOT_ACCESSIBLE
                message: This survey is not accessible in this context.
            member_self_action:
              summary: MemberSelfActionError
              value:
                code: MEMBER_SELF_ACTION
                message: You cannot modify your own membership.
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            link_not_found:
              summary: LinkNotFoundError
              value:
                code: NOT_FOUND
                message: Invalid or unknown token
            public_link_not_found:
              summary: PublicLinkNotFoundError
              value:
                code: NOT_FOUND
                message: Survey link not found
            survey_not_found:
              summary: SurveyNotFoundError
              value:
                code: SURVEY_NOT_FOUND
                message: Survey 0 was not found in project 0.
            survey_not_published:
              summary: SurveyNotPublishedError
              value:
                code: SURVEY_NOT_PUBLISHED
                message: Survey 0 in project 0 is not published.
            survey_not_found_by_slug:
              summary: SurveyNotFoundBySlugError
              value:
                code: NOT_FOUND
                message: Survey not found
            project_not_found:
              summary: ProjectNotFoundError
              value:
                code: NOT_FOUND
                message: Project id=None not found.
            version_not_found:
              summary: VersionNotFoundError
              value:
                code: NOT_FOUND
                message: Version 0 was not found in survey 0.
            node_not_found:
              summary: NodeNotFoundError
              value:
                code: NOT_FOUND
                message: Node not found
            scoring_rule_not_found:
              summary: ScoringRuleNotFoundError
              value:
                code: NOT_FOUND
                message: Scoring rule not found
            submission_not_found:
              summary: SubmissionNotFoundError
              value:
                code: SUBMISSION_NOT_FOUND
                message: Submission 0 not found.
            user_not_found:
              summary: UserNotFoundError
              value:
                code: USER_NOT_FOUND
                message: User <example> not found.
            invitation_not_found:
              summary: InvitationNotFoundError
              value:
                code: NOT_FOUND
                message: Invitation not found.
            project_role_not_found:
              summary: ProjectRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Project role not found.
            member_not_found:
              summary: MemberNotFoundError
              value:
                code: NOT_FOUND
                message: Member not found.
            survey_role_not_found:
              summary: SurveyRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Survey role not found.
            survey_member_role_not_found:
              summary: SurveyMemberRoleNotFoundError
              value:
                code: NOT_FOUND
                message: Survey member role assignment not found.
...
    ValidationError:
      description: Request was syntactically valid but semantically invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            link_auth_assignment_required:
              summary: LinkAuthAssignmentRequiredError
              value:
                code: LINK_ASSIGNED_EMAIL_REQUIRED
                message: Links that require authentication must be assigned to an email.
            private_survey_assigned_email_required:
              summary: PrivateSurveyAssignedEmailRequiredError
              value:
                code: ASSIGNED_EMAIL_REQUIRED
                message: Private surveys require links assigned to a specific email.
            survey_visibility_mismatch:
              summary: SurveyVisibilityMismatchError
              value:
                code: SURVEY_VISIBILITY_MISMATCH
                message: <example>
            ck_survey_roles_description_len:
              summary: 'DB rule: ROLE_DESCRIPTION_INVALID'
              value:
                code: ROLE_DESCRIPTION_INVALID
                message: Survey role description for project id=<project_id> is blank or exceeds 500 characters.
            ck_project_roles_description_len:
              summary: 'DB rule: ROLE_DESCRIPTION_INVALID'
              value:
                code: ROLE_DESCRIPTION_INVALID
                message: Project role description for project id=<project_id> is blank or exceeds 500 characters.
            pydantic_field_validation:
              summary: Pydantic field validation failure
              value:
                code: VALIDATION_ERROR
                message: Request validation failed.
                details:
                  errors:
                  - message: Input should be a valid integer, unable to parse string as an integer
                    type: int_parsing
                    field: sort_key
                  - message: Unable to extract tag using discriminator 'family'
                    type: union_tag_not_found
                    field: content
    RateLimitError:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit_exceeded:
              summary: RateLimitExceededError
              value:
                code: RATE_LIMIT_EXCEEDED
                message: Rate limit exceeded. Please retry after some time.
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unhandled_db_integrity:
              summary: UnhandledDbIntegrityError
              value:
                code: UNHANDLED_DB_INTEGRITY_ERROR
                message: An unexpected database integrity error occurred.
                details:
                  constraint_name: null
                  error_key: null
                  context: {}
            ck_project_invitations_status_valid:
              summary: 'DB rule: INVITATION_STATE_INVALID'
              value:
                code: INVITATION_STATE_INVALID
                message: 'Server invariant violated: invitation id=<invitation_id> was written with invalid status=''<status>''.'
            ck_survey_versions_published_requires_schema_and_timestamp:
              summary: 'DB rule: VERSION_PUBLISH_STATE_INVALID'
              value:
                code: VERSION_PUBLISH_STATE_INVALID
                message: 'Server invariant violated: survey version id=<survey_version_id> reached status=''published'' without
                  both compiled_schema and published_at. This indicates a code path bypassed surveys_repo.publish_version.'
            ck_survey_submissions_status_valid:
              summary: 'DB rule: SUBMISSION_STATE_INVALID'
              value:
                code: SUBMISSION_STATE_INVALID
                message: 'Server invariant violated: submission id=<submission_id> was written with invalid status=''<status>''.
                  Status is server-controlled and must be one of ''pending'', ''stored'', ''failed''.'
            ck_project_memberships_status_valid:
              summary: 'DB rule: MEMBERSHIP_STATE_INVALID'
              value:
                code: MEMBERSHIP_STATE_INVALID
                message: 'Server invariant violated: membership for user id=<user_id> in project id=<project_id> was written
                  with invalid status=''<status>''.'
            internal_server_error:
              summary: Generic server error
              value:
                code: INTERNAL_SERVER_ERROR
                message: An unexpected error occurred.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Auth0-issued JWT access token. Send as `Authorization: Bearer <token>`.'